Moved code and refactoring
